+++ src/main/java/cotato/growingpain/post/repository/PostSaveRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostSaveRepository extends JpaRepository<PostSave, Long> {
 
+    boolean existsById(Long postSaveId);
     boolean existsByMemberIdAndPostId(Long memberId, Long postId);
     void deleteByMemberIdAndPostId(Long memberId, Long postId);
 }
